refactor: 에디터 연산 함수 useEditorOperation 커스텀 훅으로 분리
